feat: 32-bit index migration, AOLibre map support, and profile-based storage - Refactored core data structures (GrhInfo, MapData, AssetRegistry) to use 'int' instead of 'short' for indices, supporting GRH IDs beyond 32767. - Implemented AOLibre (Version 136) map loading with 4-byte 'Long' index support. - Fixed binary alignment in MapManager by ensuring particles are always read as 2 bytes (VB6 Integer) to prevent stream desynchronization. - Migrated minimap data storage (minimap.bin) to the 'profiles/' directory for better user data isolation. - Added dynamic memory resizing for AssetRegistry.grhData to handle large, sparse GRH indices. - Improved rendering robustness in Drawn.java with better boundary checks for missing or out-of-range assets. - Updated editor tools (Surface, Npc, Transfer, AutoTiler) to maintain compatibility with the new 32-bit addressing.

+++ src/main/java/org/argentumforge/engine/utils/editor/AutoTiler.java
@@ -18,16 +18,16 @@ public class AutoTiler {
      * @param landGrh       GRH index for land tiles
      * @param coastGrhStart Starting GRH index for coast tiles (assumes 16-tile set)
      */
-    public static void applyCoasting(int layer, short waterGrh, short landGrh, short coastGrhStart) {
+    public static void applyCoasting(int layer, int waterGrh, int landGrh, int coastGrhStart) {
         if (GameData.mapData == null)
             return;
 
-        java.util.Map<org.argentumforge.engine.utils.editor.commands.BulkTileChangeCommand.TilePos, Short> oldTiles = new java.util.HashMap<>();
-        java.util.Map<org.argentumforge.engine.utils.editor.commands.BulkTileChangeCommand.TilePos, Short> newTiles = new java.util.HashMap<>();
+        java.util.Map<org.argentumforge.engine.utils.editor.commands.BulkTileChangeCommand.TilePos, Integer> oldTiles = new java.util.HashMap<>();
+        java.util.Map<org.argentumforge.engine.utils.editor.commands.BulkTileChangeCommand.TilePos, Integer> newTiles = new java.util.HashMap<>();
 
         for (int x = Camera.XMinMapSize; x <= Camera.XMaxMapSize; x++) {
             for (int y = Camera.YMinMapSize; y <= Camera.YMaxMapSize; y++) {
-                short currentGrh = GameData.mapData[x][y].getLayer(layer).getGrhIndex();
+                int currentGrh = GameData.mapData[x][y].getLayer(layer).getGrhIndex();
 
                 // Only process land tiles that border water
                 if (currentGrh == landGrh) {
@@ -35,7 +35,7 @@ public class AutoTiler {
 
                     if (bitmask > 0) {
                         // This land tile borders water, apply coast tile
-                        short coastTile = getCoastTile(coastGrhStart, bitmask);
+                        int coastTile = getCoastTile(coastGrhStart, bitmask);
                         oldTiles.put(
                                 new org.argentumforge.engine.utils.editor.commands.BulkTileChangeCommand.TilePos(x, y),
                                 currentGrh);
@@ -61,7 +61,7 @@ public class AutoTiler {
      * Bit 2 (4) = South has water
      * Bit 3 (8) = West has water
      */
-    private static int calculateBitmask(int x, int y, int layer, short waterGrh) {
+    private static int calculateBitmask(int x, int y, int layer, int waterGrh) {
         int mask = 0;
 
         // North
@@ -91,41 +91,41 @@ public class AutoTiler {
      * 4-7: Corners (NE, SE, SW, NW)
      * 8-15: Complex transitions
      */
-    private static short getCoastTile(short baseGrh, int bitmask) {
+    private static int getCoastTile(int baseGrh, int bitmask) {
         // Simplified mapping - you may need to adjust based on your tileset
         switch (bitmask) {
             case 1:
-                return (short) (baseGrh + 0); // North edge
+                return (baseGrh + 0); // North edge
             case 2:
-                return (short) (baseGrh + 1); // East edge
+                return (baseGrh + 1); // East edge
             case 4:
-                return (short) (baseGrh + 2); // South edge
+                return (baseGrh + 2); // South edge
             case 8:
-                return (short) (baseGrh + 3); // West edge
+                return (baseGrh + 3); // West edge
             case 3:
-                return (short) (baseGrh + 4); // NE corner
+                return (baseGrh + 4); // NE corner
             case 6:
-                return (short) (baseGrh + 5); // SE corner
+                return (baseGrh + 5); // SE corner
             case 12:
-                return (short) (baseGrh + 6); // SW corner
+                return (baseGrh + 6); // SW corner
             case 9:
-                return (short) (baseGrh + 7); // NW corner
+                return (baseGrh + 7); // NW corner
             case 5:
-                return (short) (baseGrh + 8); // N+S edges
+                return (baseGrh + 8); // N+S edges
             case 10:
-                return (short) (baseGrh + 9); // E+W edges
+                return (baseGrh + 9); // E+W edges
             case 7:
-                return (short) (baseGrh + 10); // N+E+S
+                return (baseGrh + 10); // N+E+S
             case 14:
-                return (short) (baseGrh + 11); // E+S+W
+                return (baseGrh + 11); // E+S+W
             case 13:
-                return (short) (baseGrh + 12); // S+W+N
+                return (baseGrh + 12); // S+W+N
             case 11:
-                return (short) (baseGrh + 13); // W+N+E
+                return (baseGrh + 13); // W+N+E
             case 15:
-                return (short) (baseGrh + 14); // All sides (island)
+                return (baseGrh + 14); // All sides (island)
             default:
-                return (short) (baseGrh + 15); // Fallback
+                return (baseGrh + 15); // Fallback
         }
     }
 
@@ -137,22 +137,22 @@ public class AutoTiler {
      * @param mosaicWidth  Width of mosaic pattern
      * @param mosaicHeight Height of mosaic pattern
      */
-    public static void applyMosaic(int layer, short baseGrh, int mosaicWidth, int mosaicHeight) {
+    public static void applyMosaic(int layer, int baseGrh, int mosaicWidth, int mosaicHeight) {
         if (GameData.mapData == null || mosaicWidth <= 1 || mosaicHeight <= 1)
             return;
 
-        java.util.Map<org.argentumforge.engine.utils.editor.commands.BulkTileChangeCommand.TilePos, Short> oldTiles = new java.util.HashMap<>();
-        java.util.Map<org.argentumforge.engine.utils.editor.commands.BulkTileChangeCommand.TilePos, Short> newTiles = new java.util.HashMap<>();
+        java.util.Map<org.argentumforge.engine.utils.editor.commands.BulkTileChangeCommand.TilePos, Integer> oldTiles = new java.util.HashMap<>();
+        java.util.Map<org.argentumforge.engine.utils.editor.commands.BulkTileChangeCommand.TilePos, Integer> newTiles = new java.util.HashMap<>();
 
         for (int x = Camera.XMinMapSize; x <= Camera.XMaxMapSize; x++) {
             for (int y = Camera.YMinMapSize; y <= Camera.YMaxMapSize; y++) {
-                short currentGrh = GameData.mapData[x][y].getLayer(layer).getGrhIndex();
+                int currentGrh = GameData.mapData[x][y].getLayer(layer).getGrhIndex();
 
                 if (currentGrh == baseGrh) {
                     // Apply tiling pattern
                     int relX = x % mosaicWidth;
                     int relY = y % mosaicHeight;
-                    short mosaicGrh = (short) (baseGrh + (relY * mosaicWidth) + relX);
+                    int mosaicGrh = (baseGrh + (relY * mosaicWidth) + relX);
 
                     if (currentGrh != mosaicGrh) {
                         oldTiles.put(
